sadim_network/requirements.txt
# Core Framework
Flask==3.1.2
Werkzeug==3.1.4

# Database
psycopg2-binary==2.9.11

# Caching & Sessions
redis==7.1.0

# Security & Authentication
Flask-Login==0.6.3
Flask-Limiter==4.1.1

# Environment Variables
python-dotenv==1.2.1

# Production Server
gunicorn==23.0.0
gevent==24.11.1

# Telegram Bot
pyTelegramBotAPI==4.14.0

# Video Download
yt-dlp==2024.12.6

# PDF Processing
PyMuPDF==1.24.1
arabic-reshaper>=3.0.0

python-bidi==0.4.2

# Utilities
pytz==2025.2
